Move Fib comp to pages folder
