+++ client/src/App.js
@@ -1,6 +1,6 @@
 import { Fragment } from "react";
 import "./App.css";
-import Fib from "./Components/Fib";
+import Fib from "./Components/Pages/Fib";
 import { Link } from "react-router-dom";
 
 function App() {
